+++ admin/src/main/scala/com/blinkbox/books/credit/admin/RenderingDtos.scala
@@ -4,9 +4,9 @@ import org.joda.time.DateTime
 
 sealed trait RenderingCreditOrDebit
 case class CreditIssuerForRendering(name: String, roles: Set[String])
-case class CreditForRendering(transactionId: String,dateTime: DateTime, amount: Money, reason: String, issuer: Option[CreditIssuerForRendering]) extends RenderingCreditOrDebit
-case class DebitForRendering(transactionId: String, dateTime: DateTime, amount: Money) extends RenderingCreditOrDebit
-case class CreditHistoryForRendering(balance: Money, items: List[RenderingCreditOrDebit])
+case class CreditForRendering(transactionId: String,dateTime: DateTime, amount: Amount, reason: String, issuer: Option[CreditIssuerForRendering]) extends RenderingCreditOrDebit
+case class DebitForRendering(transactionId: String, dateTime: DateTime, amount: Amount) extends RenderingCreditOrDebit
+case class CreditHistoryForRendering(balance: Amount, items: List[RenderingCreditOrDebit])
 
 object RenderingFunctions {
   def removeIssuer(cd: CreditOrDebit): RenderingCreditOrDebit = cd match {