+++ admin/src/test/scala/com/blinkbox/books/credit/admin/AdminApiTest.scala
@@ -8,13 +8,8 @@ import org.json4s.JsonAST.JObject
 import org.json4s._
 import org.scalatest.{BeforeAndAfter, FlatSpec}
 import org.mockito.Mockito._
-import spray.http.HttpHeaders.Authorization
-import spray.http.{ OAuth2BearerToken, StatusCodes }
-import spray.routing.authentication.{ContextAuthenticator, Authentication}
-import spray.routing._
-import spray.routing.AuthenticationFailedRejection.CredentialsRejected
-import spray.routing.authentication.{ ContextAuthenticator, Authentication }
-import spray.routing.{ Route, AuthenticationFailedRejection, RequestContext, HttpService }
+import spray.http.StatusCodes
+import spray.routing.{AuthenticationFailedRejection, RequestContext, HttpService }
 import com.blinkbox.books.spray.v2.`application/vnd.blinkbox.books.v2+json`
 import com.blinkbox.books.spray.BearerTokenAuthenticator
 import spray.testkit.ScalatestRouteTest
@@ -107,18 +102,18 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
   it should "204 on add debit endpoint, as Customer Service Manager and Customer Service Representative" in new TestFixture {
     Set(authenticatedUserCSR, authenticatedUserCSM).foreach { adminUser =>
       when(authenticator.apply(any[RequestContext])).thenReturn(Future.successful(Right(adminUser)))
-      val amount = Money(BigDecimal.valueOf(90.01), "GBP")
+      val amount = Amount(BigDecimal.valueOf(90.01), "GBP")
       when(adminService.addDebit(123, amount, "good")).thenReturn(Future.successful(()))
       Post("/admin/users/123/accountcredit/debits", debitRequest) ~> route ~> check {       
         assert(status == StatusCodes.NoContent)
       }      
     }
-    verify(adminService, atLeast(2)).addDebit(123, Money(BigDecimal.valueOf(90.01), "GBP"), "good")
+    verify(adminService, atLeast(2)).addDebit(123, Amount(BigDecimal.valueOf(90.01), "GBP"), "good")
   }
 
   it should "400 on add debit endpoint, if trying to debit more credit than they have" in new TestFixture {
     when(authenticator.apply(any[RequestContext])).thenReturn(Future.successful(Right(authenticatedUserCSR)))
-    when(adminService.addDebit(any[Int], any[Money], any[String])).thenReturn(Future.failed(new InsufficientFundsException))
+    when(adminService.addDebit(any[Int], any[Amount], any[String])).thenReturn(Future.failed(new InsufficientFundsException))
     Post("/admin/users/123/accountcredit/debits", debitRequest) ~> route ~> check {
       assert(status == StatusCodes.BadRequest)
       assert(responseAs[JObject] == errorMessage("InsufficientFunds"))
@@ -147,7 +142,7 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
     when(authenticator.apply(any[RequestContext])).thenReturn(Future.successful(Right(authenticatedUserCSR)))
     when(adminService.hasRequestAlreadyBeenProcessed("good")).thenReturn(Future.successful(true))
     Post("/admin/users/123/accountcredit/debits", debitRequest) ~> route ~> check {
-      verify(adminService, never()).addDebit(any[Int], any[Money], any[String])
+      verify(adminService, never()).addDebit(any[Int], any[Amount], any[String])
       assert(status == StatusCodes.NoContent)
     }
   }
@@ -156,7 +151,7 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
     Set("unsupported_currency", "invalid_amount").foreach { invalidRequestMsg =>
 
       when(authenticator.apply(any[RequestContext])).thenReturn(Future.successful(Right(authenticatedUserCSR)))
-      when(adminService.addDebit(any[Int], any[Money], any[String])).thenReturn(Future.failed(new InvalidRequestException(invalidRequestMsg)))
+      when(adminService.addDebit(any[Int], any[Amount], any[String])).thenReturn(Future.failed(new InvalidRequestException(invalidRequestMsg)))
 
       Post("/admin/users/12/accountcredit/debits", NewDebit(null, "")) ~> route ~> check {
         assert(status == StatusCodes.BadRequest)
@@ -199,7 +194,7 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
 
   it should "204 NoContent on add Credit with Reason, as Customer Service Manager and Customer Service Representative" in new TestFixture {
     Set(authenticatedUserCSR, authenticatedUserCSM).foreach { adminUser =>
-      val amount = Money(BigDecimal.valueOf(90.01), "GBP")
+      val amount = Amount(BigDecimal.valueOf(90.01), "GBP")
       val creditRequest = NewCredit(amount, "tests125455", "CreditVoucherCode")
 
       when(authenticator.apply(any[RequestContext])).thenReturn(Future.successful(Right(adminUser)))
@@ -241,8 +236,8 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
     }
   }
 
-  it should "204 and igore already existing request Id" in new TestFixture {
-    val amount = Money(BigDecimal.valueOf(90.01), "GBP")
+  it should "204 and ignore already existing request Id" in new TestFixture {
+    val amount = Amount(BigDecimal.valueOf(90.01), "GBP")
     val creditRequest = NewCredit(amount, "alreadyExistRequestId", "CreditVoucherCode")
 
     when(authenticator.apply(any[RequestContext])).thenReturn(Future.successful(Right(authenticatedUserCSR)))
@@ -263,7 +258,7 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
   it should "should throw InvalidRequestException on invalid amount" in new TestFixture {
     Set(0, -1).foreach { wrongValue =>
       val e = intercept[InvalidRequestException] {
-        defaultAdminService.validateAmount(Money(BigDecimal(wrongValue), "GBP"))
+        defaultAdminService.validateAmount(Amount(BigDecimal(wrongValue)))
       }
       assert(e.message == "invalid_amount")
     }
@@ -271,7 +266,7 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
 
   it should "should throw InvalidRequestException on unsupported currency" in new TestFixture {
     val e = intercept[InvalidRequestException] {
-      defaultAdminService.validateAmount(Money(BigDecimal(1), "USD"))
+      Amount(BigDecimal(1), "USD")
     }
     assert(e.message == "unsupported_currency")
   }
@@ -303,7 +298,7 @@ class AdminApiTest extends FlatSpec with ScalatestRouteTest with HttpService wit
 object AdminApiTest {
   val dummy = {
   val thePast = new DateTime(2012,1,2,3,4,5)
-  val cheap = Money(BigDecimal.valueOf(1000.53))
+  val cheap = Amount(BigDecimal(1000.53))
   val requestId = "sdfnaksfniofgniaodoir84t839t"
   val credits = List(Credit(requestId, thePast, cheap, "CreditVoucherCode", CreditIssuer("James Bond", Set(UserRole.CustomerServicesRep))))
   val debits = List(Debit(requestId, thePast, cheap))
@@ -315,7 +310,7 @@ object AdminApiTest {
   CreditHistory(cheap, eithers)
 }
 
-  val zeroCreditHistory = CreditHistory(Money(BigDecimal(0)), List())
+  val zeroCreditHistory = CreditHistory(Amount(BigDecimal(0)), List())
 }
 
 class TestFixture extends v2.JsonSupport with MockitoSyrup  {
@@ -339,9 +334,8 @@ class TestFixture extends v2.JsonSupport with MockitoSyrup  {
 
   when(authenticator.withElevation(Elevation.Critical)).thenReturn(authenticator)
 
-  val debitRequest = NewDebit(Money(BigDecimal.valueOf(90.01), "GBP"), "good")
-  val nonGbpDebitRequest = NewDebit(Money(BigDecimal.valueOf(90.01), "USD"), "good")
-  
+  val debitRequest = NewDebit(Amount(BigDecimal.valueOf(90.01), "GBP"), "good")
+
   val accountCreditStore = mock[AccountCreditStore]
   val clock = StoppedClock()
   val defaultAdminService = new DefaultAdminService(accountCreditStore, clock)